Move new impl to same package as others.
Also clean up
